Refactor : organize box data mapper
