+++ src/dataMappers/boxDataMapper/index.js
@@ -1,0 +1,15 @@
+const getBoxByIdDM = require('./getBoxById');
+const getBoxesDM = require('./getBoxes');
+const createBoxDM = require('./createBox');
+const updateBoxDM = require('./updateBox');
+const updateBoxLearnItValueDM = require('./updateBoxLearnItValue');
+const deleteBoxDM = require('./deleteBox');
+
+module.exports = {
+  getBoxByIdDM,
+  getBoxesDM,
+  createBoxDM,
+  updateBoxDM,
+  updateBoxLearnItValueDM,
+  deleteBoxDM,
+};
