+++ src/dataMappers/cardDataMapper.js
@@ -1,4 +1,4 @@
-const { pool } = require('../database');
+const { pool } = require('./database');
 
 async function getCardById(id) {
   try {
