+++ src/dataMappers/statsDataMapper.js
@@ -1,4 +1,4 @@
-const { pool } = require('../database');
+const { pool } = require('./database');
 
 async function getInstantStats(boxId) {
   try {
